fix(bundling): Extract getAllToyENSEntries from node
It is used from mangrove.ts and thus from browsers.
Browsers cannot access node.
